characteristic/u2b18/src/main/java/org/im97mori/ble/characteristic/u2b18/VoltageAndroid.java
package org.im97mori.ble.characteristic.u2b18;

import android.bluetooth.BluetoothGattCharacteristic;
import android.os.Parcel;
import android.os.Parcelable;

import androidx.annotation.NonNull;

import org.im97mori.ble.ByteArrayCreator;

import java.util.Objects;

/**
 * Voltage (Characteristics UUID: 0x2B18)
 */
@SuppressWarnings({"WeakerAccess"})
public class VoltageAndroid extends Voltage implements Parcelable {

    /**
     * @see ByteArrayCreator
     */
    public static final ByteArrayCreator<VoltageAndroid> CREATOR = new ByteArrayCreator<VoltageAndroid>() {

        /**
         * {@inheritDoc}
         */
        @Override
        @NonNull
        public VoltageAndroid createFromParcel(@NonNull Parcel in) {
            return new VoltageAndroid(in);
        }

        /**
         * {@inheritDoc}
         */
        @Override
        @NonNull
        public VoltageAndroid[] newArray(int size) {
            return new VoltageAndroid[size];
        }

        /**
         * {@inheritDoc}
         */
        @NonNull
        public VoltageAndroid createFromByteArray(@NonNull byte[] values) {
            return new VoltageAndroid(values);
        }

    };

    /**
     * Constructor from {@link BluetoothGattCharacteristic}
     *
     * @param bluetoothGattCharacteristic Characteristics UUID: 0x2B18
     */
    @Deprecated
    public VoltageAndroid(@NonNull BluetoothGattCharacteristic bluetoothGattCharacteristic) {
        super(bluetoothGattCharacteristic.getValue());
    }

    /**
     * Constructor from byte array
     *
     * @param values byte array from <a href="https://developer.android.com/reference/android/bluetooth/BluetoothGattCharacteristic#getValue()">BluetoothGattCharacteristic#getValue()</a>
     */
    public VoltageAndroid(@NonNull byte[] values) {
        super(values);
    }

    /**
     * Constructor from parameters
     *
     * @param voltageValue Voltage Value
     */
    public VoltageAndroid(int voltageValue) {
        super(voltageValue);
    }

    /**
     * Constructor from {@link Parcel}
     *
     * @param in Parcel
     */
    private VoltageAndroid(@NonNull Parcel in) {
        super(Objects.requireNonNull(in.createByteArray()));
    }

    /**
     * {@inheritDoc}
     */
    @Override
    public int describeContents() {
        return 0;
    }

    /**
     * {@inheritDoc}
     */
    @Override
    public void writeToParcel(@NonNull Parcel dest, int flags) {
        dest.writeByteArray(getBytes());
    }

}
